requirements.txt
asgiref==3.11.1
dj-database-url==3.1.0
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.3
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.11.0
